fix:リツイート時の不具合を修正
リツイートをリツイートした時に元の投稿の画像と名前が表示されない不具合を修正しました

+++ html/php-challenge/good.php
@@ -2,62 +2,63 @@
 session_start();
 require('dbconnect.php');
 
-// 投稿を検査 
+// 投稿をカウント いいねの有無の確認に使用
 $goodmessages = $db->prepare('SELECT count(*) FROM good WHERE post_id=? AND member_id=?');
 $goodmessages->execute(array(
-  $_POST['post_id'],
+  $_POST['original_post_id'],
   $_SESSION['id']
 ));
 $goodmessage = $goodmessages->fetch();
 
 
-
-// $_POST['reference']に値があればリツイートの投稿だとわかる
-
-$iineids = $db->prepare('SELECT * FROM good WHERE post_id=? and member_id=?');
-$iineids->execute(array(
-  $_POST['reference'],
+//投稿を検索 いいねの挿入・削除の判定に使用
+$goodrecords = $db->prepare('SELECT * FROM good WHERE post_id=? AND member_id=?');
+$goodrecords->execute(array(
+  $_POST['post_id'],
   $_SESSION['id']
 ));
-$iineid = $iineids->fetch();
+$goodrecord = $goodrecords->fetch();
 
-// if ($_POST['reference'] > 0) 
-// {
-//   $rtdelete = $db->prepare('DELETE FROM good WHERE post_id=? AND member_id=?');
-//   $rtdelete->execute(array(
-//     $_POST['reference'],
-//     $_SESSION['id']
-//   ));
-// }
-// else
-if ($_POST['reference'] = 0) 
-{
-  $rtgood = $db->prepare('INSERT INTO good SET post_id=?, member_id=?, created=NOW()');
-  $rtgood->execute(array(
-    $_POST['reference'],
+// var_dump($goodmessage['count(*)']);
+// var_dump($good);
+// var_dump($_POST['original_post_id']);
+
+
+
+
+// いいねテーブルのレコードの削除・挿入
+if ($goodmessage['count(*)'] > 0) {
+  $delete = $db->prepare('DELETE FROM good WHERE post_id=? AND member_id=?');
+  $delete->execute(array(
+    $_POST['original_post_id'],
     $_SESSION['id']
   ));
+} else {
+  //リツイートにいいねした場合
+  if ($_POST['original_post_id']) {
+    $rtgood = $db->prepare('INSERT INTO good SET post_id=?, member_id=?, created=NOW()');
+    $rtgood->execute(array(
+      $_POST['original_post_id'],
+      $_SESSION['id']
+    ));
+  } else {
+    //普通メッセージにいいねした場合
+    if ($goodrecord) {
+      $delete = $db->prepare('DELETE FROM good WHERE post_id=? AND member_id=?');
+      $delete->execute(array(
+        $_POST['post_id'],
+        $_SESSION['id']
+      ));
+    } else {
+      $good = $db->prepare('INSERT INTO good SET post_id=?, member_id=?,  created=NOW()');
+      $good->execute(array(
+        $_POST['post_id'],
+        $_SESSION['id']
+      ));
+    }
+  }
 }
 
 
-// var_dump($_POST['reference']);
-
-// いいねテーブルのレコードの削除・挿入
-if ( $goodmessage['count(*)'] > 0) {
-  $delete = $db->prepare('DELETE FROM good WHERE post_id=? AND member_id=?');
-  $delete->execute(array(
-    $_POST['post_id'],
-    $_SESSION['id']
-  ));
-}else { 
-    //テーブルpostのidがテーブルgoodのpost_idにINSERTされる
-    $good = $db->prepare('INSERT INTO good SET post_id=?, member_id=?, created=NOW()');
-    $good->execute(array(
-      $_POST['post_id'],
-      $_SESSION['id']
-    ));
-  }
-
-
-header('Location: index.php'); exit();
-
+header('Location: index.php');
+exit();
